refactor: move type defs to root, range class to separate file
