+++ lib/corefoundation/memory.rb
@@ -2,25 +2,6 @@
 
 # The top level namespace for the corefoundation library. The raw FFI generated methods are attached here
 module CF
-  extend FFI::Library
-  ffi_lib '/System/Library/Frameworks/CoreFoundation.framework/CoreFoundation'
-
-  if FFI::Platform::ARCH == 'x86_64'
-    typedef :long_long, :cfindex
-    typedef :long_long, :cfcomparisonresult
-    typedef :ulong_long, :cfoptionflags
-    typedef :ulong_long, :cftypeid
-    typedef :ulong_long, :cfhashcode
-  else
-    typedef :long, :cfindex
-    typedef :long, :cfcomparisonresult
-    typedef :ulong, :cfoptionflags
-    typedef :ulong, :cftypeid
-    typedef :ulong, :cfhashcode
-  end
-
-  typedef :pointer, :cftyperef
-
   attach_function :show, 'CFShow', [:cftyperef], :void
   attach_function :release, 'CFRelease', [:cftyperef], :void
   attach_function :retain, 'CFRetain', [:cftyperef], :cftyperef
@@ -28,12 +9,6 @@ module CF
   attach_function 'CFHash', [:cftyperef], :cfhashcode
   attach_function 'CFCopyDescription', [:cftyperef], :cftyperef
   attach_function 'CFGetTypeID', [:cftyperef], :cftypeid
-
-  # @private
-  class Range < FFI::Struct
-    layout :location, :cfindex,
-           :length, :cfindex
-  end
 
   # Provides a more declarative way to define all required abstract methods.
   # This gets included in the CF::Base class whose subclasses then need to define those methods.
